Options available for McCormick Operators

Using fast vs. correctly rounded intervals

Correctly rounded versus fast intervals are selected by setting the type V, in the SMCg{N,V,T} object to either:

Interval{T} for correctly rounded calculations

MCInterval{T} for fast calculations without corrected rounding

Setting the differentiability of the McCormick relaxation

EAGO.set_diff_relax — Method.

set_diff_relax(val::Integer)

Set differentiability of relaxations used.

source

Using nonsmooth multivariate McCormick relaxations

EAGO.set_multivar_refine — Method.

set_multivar_refine(bool,tol)

Set flag for using using multivariant MC relaxations and their tolerance.

source

Using subgradient refinement

EAGO.set_subgrad_refine — Method.

set_subgrad_refine(val)

Set flag for using subgradient refinement of interval bounds.

source

Setting the tolerance used in envelope calculations

EAGO.set_tolerance — Method.

set_tolerance(val::Float64)

Set tolerance for used in envelope calculations.

source

Setting the number of iterations used for the envelope calculation

EAGO.set_iterations — Method.

set_iterations(val::Integer)

Set iterations for used in envelope calculations.

source