Relaxation of Implicit Functions

High-level functions

McCormick.implicit_relax_h! — Function

implicit_relax_h!(d)
implicit_relax_h!(d, interval_bnds)

Compute relaxations of x(p) defined by h(x,p) = 0 where h is specifed as h(out, x, p).

DataStructures

McCormick.AbstractContractorMC — Type

abstract type AbstractContractorMC

An abstract type for each manner of contractor using in the implicit function relaxation algorithms.

McCormick.NewtonGS — Type

struct NewtonGS <: McCormick.AbstractContractorMC

The Gauss-Seidel implementation of the Newton contractor used in the implicit relaxation scheme.

McCormick.KrawczykCW — Type

struct KrawczykCW <: McCormick.AbstractContractorMC

The componentwise implementation of the Krawczyk contractor used in the implicit relaxation scheme.

McCormick.AbstractPreconditionerMC — Type

abstract type AbstractPreconditionerMC

An abstract type for each manner of preconditioner used in the implicit function relaxation algorithms.

McCormick.DenseMidInv — Type

struct DenseMidInv{S<:Union{Array{Float64,1}, Array{Float64,2}}} <: McCormick.AbstractPreconditionerMC

A dense LU preconditioner for implicit McCormick relaxation.

McCormick.AbstractMCCallback — Type

abstract type AbstractMCCallback

An abstract type for each manner of callback functions used in the implicit function relaxation algorithms.

McCormick.MCCallback — Type

struct MCCallback{FH<:Function, FJ<:Function, C<:McCormick.AbstractContractorMC, PRE<:McCormick.AbstractPreconditionerMC, N, T<:RelaxTag, AMAT<:(AbstractArray{T,2} where T)} <: AbstractMCCallback

A structure used to compute implicit relaxations.

h!::Function

Function h(x,p) = 0 defined in place by h!(out,x,p)

hj!::Function

Jacobian of h(x,p) w.r.t x

H::Array{MC{N,T},1} where T<:RelaxTag where N

J::AbstractArray{T,2} where T

xmid::Array{Float64,1}

X::Array{Interval{Float64},1}

P::Array{Interval{Float64},1}

nx::Int64

State space dimension

np::Int64

Decision space dimension

lambda::Float64

Convex combination parameter

eps::Float64

Tolerance for interval equality

kmax::Int64

Number of contractor steps to take

pref_mc::Array{MC{N,T},1} where T<:RelaxTag where N

p_mc::Array{MC{N,T},1} where T<:RelaxTag where N

x0_mc::Array{MC{N,T},1} where T<:RelaxTag where N

x_mc::Array{MC{N,T},1} where T<:RelaxTag where N

xa_mc::Array{MC{N,T},1} where T<:RelaxTag where N

xA_mc::Array{MC{N,T},1} where T<:RelaxTag where N

aff_mc::Array{MC{N,T},1} where T<:RelaxTag where N

z_mc::Array{MC{N,T},1} where T<:RelaxTag where N

contractor::McCormick.AbstractContractorMC

Type of contractor used in implicit relaxation routine.

preconditioner::McCormick.AbstractPreconditionerMC

Preconditioner used in the implicit relaxation routine.

apply_precond::Bool

Boolean indicating that the preconditioner should be applied

param::Array{Array{MC{N,T},1},1} where T<:RelaxTag where N

Subroutines

McCormick.preconditioner_storage — Function

preconditioner_storage(x, t)

Creates storage corresponding to x::AbstractPreconditionerMC and t::T where T<:RelaxTag.

McCormick.affine_exp! — Function

affine_exp!(x, p, d)

Computates the affine relaxations of the state variable.

McCormick.correct_exp! — Function

Corrects the relaxation of the state variable x_mc if the affine relaxation,

McCormick.contract! — Function

Performs a single step of the parametric method associated with t assumes that the inputs have been preconditioned.

McCormick.precond_and_contract! — Function

McCormick.final_cut — Function

final_cut(x, y)

An operator that cuts the x object using the y bounds in a differentiable or nonsmooth fashion to achieve a composite relaxation within y.

McCormick.gen_expansion_params! — Function

gen_expansion_params!(d)
gen_expansion_params!(d, interval_bnds)

Constructs parameters need to compute relaxations of h.