EAGO's Branch and Bound Routine

This component is meant to provide a flexible framework for implementing spatial branch-and-bound based optimization routines in Julia. All components of the branch-and-bound routine can be customized by the individual user: lower bounding problem, upper bounding problem.

Branch and Bound Node Storage

EAGO.NodeBB — Type

struct NodeBB

Stores information associated with each node in Branch & Bound tree.

lower_variable_bounds::Array{Float64,1}

Lower bounds of variable box.

upper_variable_bounds::Array{Float64,1}

Upper bounds of variable box.

lower_bound::Float64

Lower bound of problem solution on nodeBB

upper_bound::Float64

Upper bound of problem solution on nodeBB

depth::Int64

Depth of node in B&B tree.

id::Int64

Unique id for each node.

Customizable subroutines

EAGO.add_cut! — Method

add_cut!(t, m)

Adds a cut for each constraint and the objective function to the subproblem.

EAGO.branch_node! — Method

branch_node!(t, m)

Creates two nodes from current_node using information available the x and stores them to the stack. By default, relative width bisection is perfomed at a point branch_pnt which is a convex combination (parameter: branch_cvx_factor) of the solution to the relaxation and the midpoint of the node. If this solution lies within branch_offset/width of a bound then the branch point is moved to a distance of branch_offset/width from the bound.

EAGO.convergence_check — Method

convergence_check(t, m)

Checks for convergence of algorithm with respect to absolute and/or relative tolerances.

EAGO.cut_condition — Method

cut_condition(t, m)

Checks if a cut should be added and computes a new reference point to add the cut at. If no cut should be added the constraints not modified in place are deleted from the relaxed optimizer and the solution is compared with the interval lower bound. The best lower bound is then used.

EAGO.fathom! — Method

fathom!(t, m)

Selects and deletes nodes from stack with lower bounds greater than global upper bound.

EAGO.lower_problem! — Method

lower_problem!(t, m)

Constructs and solves the relaxation using the default EAGO relaxation scheme and optimizer on node y.

EAGO.node_selection! — Method

node_selection!(t, m)

Selects node with the lowest lower bound in stack.

EAGO.optimize_hook! — Method

optimize_hook!(t, m)

Provides a hook for extensions to EAGO as opposed to standard global, local, or linear solvers.

EAGO.postprocess! — Method

postprocess!(t, m)

Default postprocess perfoms duality-based bound tightening on the y.

EAGO.preprocess! — Method

preprocess!(t, m)

Runs interval, linear, quadratic contractor methods followed by obbt and a constraint programming walk up to tolerances specified in EAGO.Optimizer object.

EAGO.repeat_check — Method

repeat_check(t, m)

Checks to see if current node should be reprocessed.

EAGO.single_storage! — Method

single_storage!(t, m)

Stores the current node to the stack after updating lower/upper bounds.

EAGO.termination_check — Method

termination_check(t, m)

Checks for termination of algorithm due to satisfying absolute or relative tolerance, infeasibility, or a specified limit, returns a boolean valued true if algorithm should continue.

EAGO.upper_problem! — Method

upper_problem!(t, m)

Default upper bounding problem which simply calls solve_local_nlp! to solve the nlp locally.

Internal Subroutines

EAGO.cut_update! — Method

cut_update!(m)

Updates the internal storage in the optimizer after a valid feasible cut is added.

EAGO.interval_bound — Function

Computes a tuple representing the lower and upper interval bounds for a AbstractEAGOConstraint representing an equality constraint.

EAGO.lower_interval_bound — Function

Computes the lower interval bound for AbstractEAGOConstraint representing an inequality constraint.

EAGO.is_globally_optimal — Method

is_globally_optimal(t, r)

Takes an MOI.TerminationStatusCode and a MOI.ResultStatusCode and returns the tuple (valid_result::Bool, feasible::Bool). The value valid_result is true if the pair of codes prove that either the subproblem solution was solved to global optimality or the subproblem solution is infeasible. The value of feasible is true if the problem is feasible and false if the problem is infeasible.

EAGO.is_feasible_solution — Method

is_feasible_solution(t, r)

Takes an MOI.TerminationStatusCode and a MOI.ResultStatusCode and returns true if this corresponds to a solution that is proven to be feasible. Returns false otherwise.

EAGO.log_iteration! — Method

log_iteration!(x::Optimizer)

If 'loggingon' is true, the 'globallowerbound', 'globalupperbound', 'runtime', and 'nodecount' are stored every 'loginterval'. If 'logsubprobleminfo' then the lower bound, feasibility and run times of the subproblems are logged every 'log_interval'.

EAGO.same_box — Method

same_box

Checks that node x and y have equal domains withing a tolerance of atol.

EAGO.solve_local_nlp! — Method

Constructs and solves the problem locally on on node y updated the upper solution informaton in the optimizer.

EAGO.set_dual! — Method

set_dual!(m)

Retrieves the lower and upper duals for variable bounds from the relaxed_optimizer and sets the appropriate values in the _lower_lvd and _lower_uvd storage fields.

EAGO.update_relaxed_problem_box! — Function

update_relaxed_problem_box!(m)

Updates the relaxed constraint by setting the constraint set of v == x*,xLi <= xi, andxi <= xUi` for each such constraint added to the relaxed optimizer.

Functions for generating console output

EAGO.print_iteration! — Function

print_iteration!

Prints the iteration information based on verbosity. The header is displayed every header_interval, the iteration info is displayed every iteration_interval.

EAGO.print_node! — Function

print_node!

Prints node information for the B&B problem. Node id, bound, and interval box.

EAGO.print_results! — Function

print_results!

Prints the results of a single bounding problem.

EAGO.print_solution! — Function

print_solution!

Prints solution information for the B&B problem. Displays first node found, solution value, solution, and time spent solving subproblems.

Interval Representations of Expressions

EAGO.AbstractEAGOConstraint — Type

abstract type AbstractEAGOConstraint

An abstract super-type used for representing constraints built by EAGO's backend.

EAGO.AffineFunctionEq — Type

mutable struct AffineFunctionEq <: EAGO.AbstractEAGOConstraint

Current only used for bound tightening. Stores a representation of an affine equality.

EAGO.AffineFunctionIneq — Type

mutable struct AffineFunctionIneq <: EAGO.AbstractEAGOConstraint

Current only used for bound tightening. Stores a representation of an affine inequality.

EAGO.BufferedQuadraticIneq — Type

mutable struct BufferedQuadraticIneq <: EAGO.AbstractEAGOConstraint

Stores a general quadratic inequality constraint with a buffer.

EAGO.BufferedQuadraticEq — Type

mutable struct BufferedQuadraticEq <: EAGO.AbstractEAGOConstraint

Stores a general quadratic equality constraint with a buffer.

EAGO.NonlinearExpression — Type

mutable struct NonlinearExpression{V} <: EAGO.AbstractEAGOConstraint

Stores a general quadratic function with a buffer.

EAGO.BufferedNonlinearFunction — Type

mutable struct BufferedNonlinearFunction{V} <: EAGO.AbstractEAGOConstraint

Stores a general nonlinear function with a buffer represented by the sum of a tape and a scalar affine function.