Relaxation Backend

Nonlinear Evaluation

EAGO makes use of a specialized tape structure for each function in order to compute valid composite bounds and relaxations. The details of this are included below:

Forward Operator Updates

Main routine

EAGO.forward_pass_kernel! — Function

forwardpasskernel!

Performs a forward pass using the tape information passed as arguments. Each variety of node calls an associated forward_xxx function where xxx is a descriptor.

Subroutines to evaluation expression

EAGO.forward_plus_narity! — Function

forwardplusnarity!

Updates storage tapes with forward evalution of node representing n = +(x, y, z,...).

EAGO.forward_plus_binary! — Function

forwardplusbinary!

Updates storage tapes with forward evalution of node representing n = x + y.

EAGO.forward_minus! — Function

forward_minus!

Updates storage tapes with forward evalution for node representing n = x-y.

EAGO.forward_multiply_binary! — Function

forwardmultiplybinary!

Updates storage tapes with forward evalution for node representing n = x*y.

EAGO.forward_multiply_narity! — Function

forwardmultiplynarity!

Updates storage tapes with forward evalution of node representing n = *(x, y, z,...).

EAGO.forward_divide! — Function

forward_divide!

Updates storage tapes with forward evalution for node representing n = x/y.

EAGO.forward_power! — Function

forward_power!

Updates storage tapes with forward evalution for node representing n = x^y.

EAGO.forward_univariate_user! — Function

forwardunivariateuser!

Updates storage tapes with forward evalution for node representing n = user_f(x).

EAGO.forward_univariate_number! — Function

forwardunivariatenumber!

Updates storage tapes with forward evalution for node representing n = f(c) where f is standard function and c is a number.

EAGO.forward_univariate_tiepnt_1! — Function

forwardunivariatetiepnt_1!

Updates storage tapes with forward evalution for node representing n = f(x) where f is standard function that requires a single tiepoint calculation per convex/concave relaxation (e.g. tan).

EAGO.forward_univariate_tiepnt_2! — Function

forwardunivariatetiepnt_2!

Updates storage tapes with forward evalution for node representing n = f(x) where f is standard function that requires a two tiepoint calculations per convex/concave relaxation (e.g. sin).

EAGO.forward_univariate_other! — Function

forwardunivariateother!

Updates storage tapes with forward evalution for node representing n = f(x) where f is a standard function that does not require a tiepoint evaluation (e.g. exp).

EAGO.forward_user_multivariate! — Function

forwardusermultivariate!

Updates storage tapes with forward evalution for node representing n = user_f(x, y...).

EAGO.forward_get_subexpression! — Function

forwardgetsubexpression!

Unpacks the MC{N1,T} in the subexpression to a MC{N2,T} where N1 is the sparsity of the subexpression and N2 is the sparsity of the function tape. Note that the sparsity of the subexpression is less than the sparsity of the function itself. This limits the storage required by tapes but prevents reverse mode subgradient propagation through expressions [This may be subject to change in the future once the reverse mode propagation becomes more robust].

Reverse Operator Updates

Main routine

EAGO.reverse_pass! — Function

reverse_pass!

A reverse_pass! on a BufferedNonlinear structure d intersects the existing value of the d with constraint bounds then reverse propagates a set-valued operator (by default McCormick operator) along the computational tape. The tapes are updated in place and boolean value is returned indicating whether the reverse propagation yeilded a infeasible point (true = still feasible, false is proved infeasible).

EAGO.reverse_pass_kernel! — Function

Performs a reverse McCormick/interval pass. If a NaN value is computed for the McCormick relaxation then the routine defaults to the interval value instead.

There is a tacit assumption here that an abstract tree structure is used for propagation. If a cse structure is eventually used then each reverse_xxx function will need to keep track of each variables state.

Subroutines to evaluation expression

This list is not yet complete future documentation to be added.

EAGO.reverse_plus_binary! — Function

reverseplusbinary!

Updates storage tapes with reverse evalution of node representing n = x + y which updates x & y.

EAGO.reverse_plus_narity! — Function

reverseplusnarity!

Updates storage tapes with reverse evalution of node representing n = +(x,y,z...) which updates x, y, z and so on.

EAGO.reverse_multiply_binary! — Function

reversemultiplybinary!

Updates storage tapes with reverse evalution of node representing n = x * y which updates x & y.

EAGO.reverse_multiply_narity! — Function

reversemultiplynarity!

Updates storage tapes with reverse evalution of node representing n = *(x,y,z...) which updates x, y, z and so on.

Other routines

EAGO.set_value_post — Function

setvaluepost

Post process set_value operator. By default, performs the affine interval cut on a MC structure.

EAGO.overwrite_or_intersect — Function

overwriteorintersect

Intersects the new set valued operator with the prior and performs affine bound tightening

First forward pass: is_post should be set by user option, is_intersect should be false so that the tape overwrites existing values, and the interval_intersect flag could be set to either value.

Forward CP pass (assumes same reference point): is_post should be set by user option, is_intersect should be true so that the tape intersects with existing values, and the interval_intersect flag should be false.

Forward CP pass (assumes same reference point): is_post should be set by user option, is_intersect should be true so that the tape intersects with existing values, and the interval_intersect flag should be false.

Subsequent forward passes at new points: ispostshould be set by user option,isintersectshould be true so that the tape intersects with existing values, and theinterval_intersectflag should betrue` as predetermined interval bounds are valid but the prior values may correspond to different points of evaluation.